Created the Pie chart and Barchart for Sprint-Wise issues.
The Sprint-wise issues summary is visualized through a bar chart and pie chart, illustrating the distribution of issues across different sprints. These charts provide a clear overview of issue counts per sprint, aiding in project management and decision-making.
