+++ client/src/App.jsx
@@ -18,8 +18,6 @@ import Project from "./Project/Project";
 import IssueInfo from "./Project/ProjectIssues/IssueInfo/IssueInfo";
 import ProjectSettings from "./Project/ProjectSettings/ProjectSettings";
 import Workspace from "./Workspace/workspace";
-import ChartComponent from "./Project/chart_status"
-import Barchart_Mem from "./Project/barchart_member";
 import ActivePage from "./Project/ActivePage/ActivePage";
 import WorkPage from "./Create_workspace/Components/WorkPage";
 import MemberSetting from "./Workspace/Settings/Member/MemberSetting";
@@ -96,9 +94,7 @@ function App() {
         <Route path="*" element={<Error />} />
 
         <Route path="/workspace" element={<Workspace />} />
-        <Route path="/barpie" element={<ChartComponent />} />
         <Route path="/Activepage" element={<ActivePage />} />
-        <Route path="/barchart" element={<Barchart_Mem />} />
 
         <Route path="/workspace/project" element={<Project />} />
